Int conversion warnings

+++ src/native/credentials_provider.c
@@ -202,7 +202,7 @@ JNIEXPORT jlong JNICALL
         &proxy_options,
         &proxy_tls_connection_options,
         jni_proxy_host,
-        jni_proxy_port,
+        (uint16_t)jni_proxy_port,
         jni_proxy_authorization_username,
         jni_proxy_authorization_password,
         jni_proxy_authorization_type,
